Refactor: Removed dot imports and organized changes into the systemsmanager folder.
